dhcpv6: move option code types; add Stringer.
